claude-dc-implementation/computeruse/custom/requirements.txt
anthropic>=0.21.1
streamlit>=1.31.0
pillow>=10.0.0
nest-asyncio>=1.5.8
pytest>=7.4.0
pyautogui>=0.9.54